0.0.185 portal, signs, sheep, time save etc

+++ Assets/Script/Game/Brain/Machine/EntityMachine.cs
@@ -26,7 +26,7 @@ public abstract class EntityMachine : Machine, IInfoProvider
             if (Entity.Bounds != Vector3Int.zero)
             {
                 BoxCollider boxCollider = gameObject.AddComponent<BoxCollider>();
-                boxCollider.center = new Vector3(0, (float)Entity.Bounds.y / 2, 0); 
+                boxCollider.center = new Vector3(0, Entity.Bounds.y / 2, 0); 
                 boxCollider.size = Entity.Bounds; 
             } 
             OnSetup();
